Update Sum of Series.java

+++ Basic Java/Sum of Series.java
@@ -2,7 +2,7 @@
 
 class Solution {
     // function to return sum of  1, 2, ... n
-    long seriesSum(int n) {
+    long seriesSum(long n) {
         // code here
         n = (n+1)/2;
         return n;
